refactor(logger): replace custom logger implementation with winston
- Remove custom logger service and related interfaces/validators
- Implement new logger service using winston with configurable transports
- Add type definitions for winston-postgresql transport
- Update tests to reflect new logger implementation
