+++ src/infrastructure/services/logger/__tests__/logger.service.test.ts
@@ -1,156 +1,154 @@
-import { LoggerService } from "@/infrastructure/services/logger/logger.service";
+import { transports } from "winston";
 
-const mockExecute = jest.fn().mockResolvedValue(undefined);
-
-jest.mock("@/domain/log/use-cases/create-log", () => ({
-  CreateLog: jest.fn().mockImplementation(() => ({
-    execute: mockExecute,
-  })),
-}));
-
-jest.mock("@/domain/log/dtos/create-log.dto", () => {
-  const actual = jest.requireActual("@/domain/log/dtos/create-log.dto");
-  return {
-    __esModule: true,
-    ...actual,
-    CreateLogDto: {
-      ...actual.CreateLogDto,
-      createFrom: jest.fn((data: any) => [
-        undefined,
-        new actual.CreateLogDto(
-          data.level,
-          data.message,
-          data.meta,
-          data.service,
-          undefined,
-          data.requestId,
-          data.error,
-        ),
-      ]),
-    },
-  };
-});
-
-describe("LoggerService", () => {
-  const makeReq = (overrides: any = {}) => ({
-    method: "GET",
-    path: "/p",
-    url: "/p",
-    ip: "127.0.0.1",
-    httpVersion: "1.1",
-    get: jest.fn().mockReturnValue("UA"),
-    requestId: "req-1",
-    ...overrides,
-  });
-  const makeRes = (overrides: any = {}) => ({ statusCode: 200, ...overrides });
-
+describe("logger.service.test", () => {
   beforeEach(() => {
     jest.clearAllMocks();
-  });
-
-  it("logError calls CreateLog.execute in dev", () => {
-    const req = makeReq();
-    const res = makeRes();
-    LoggerService.logError({
-      error: new Error("boom"),
-      req,
-      res,
-      service: "svc",
-    });
-    expect(mockExecute).toHaveBeenCalledTimes(1);
-  });
-
-  it("logWarn calls CreateLog.execute in dev", () => {
-    const req = makeReq();
-    const res = makeRes();
-    LoggerService.logWarn({
-      error: new Error("boom"),
-      req,
-      res,
-      service: "svc",
-    });
-    expect(mockExecute).toHaveBeenCalledTimes(1);
-  });
-
-  it("logInfo calls in dev", () => {
-    const req = makeReq();
-    const res = makeRes();
-    LoggerService.logInfo({ message: "m", req, res, service: "svc" });
-    expect(mockExecute).toHaveBeenCalledTimes(1);
-  });
-
-  it("logDebug calls in dev", () => {
-    const req = makeReq();
-    LoggerService.logDebug({ message: "m", req, service: "svc" });
-    expect(mockExecute).toHaveBeenCalledTimes(1);
-  });
-
-  describe("generateLogData via public methods", () => {
-    it("composes meta from req, res and custom meta", () => {
-      const longUA = "Mozilla/5.0" + "x".repeat(300);
-      const req = makeReq({
-        get: jest.fn().mockReturnValue(longUA),
-        path: "/hello",
-        url: "/hello?x=1",
-      });
-      const res = makeRes({ statusCode: 201 });
-      LoggerService.logDebug({
-        message: "meta",
-        req,
-        res,
-        service: "svc",
-        meta: { extra: 1 },
-      });
-      const dto = (mockExecute as jest.Mock).mock.calls[0][0];
-      expect(dto.service).toBe("svc");
-      expect(dto.requestId).toBe("req-1");
-      expect(dto.message).toBeDefined();
-      expect(dto.meta).toBeDefined();
-      expect(dto.meta.method).toBe("GET");
-      expect(dto.meta.endpoint).toBe("/hello");
-      expect(dto.meta.url).toBe("/hello?x=1");
-      expect(dto.meta.ip).toBe("127.0.0.1");
-      expect(dto.meta.httpVersion).toBe("1.1");
-      expect(dto.meta.statusCode).toBe(201);
-      expect(dto.meta.extra).toBe(1);
-      expect(typeof dto.meta.userAgent).toBe("string");
-      expect(dto.meta.userAgent.length).toBeLessThanOrEqual(100);
-    });
-
-    it("maps error to string in logError", () => {
-      jest.clearAllMocks();
-      const req = makeReq();
-      const res = makeRes();
-      const err = new Error("boom");
-      LoggerService.logError({ error: err, req, res, service: "svc" });
-      const dto = (mockExecute as jest.Mock).mock.calls[0][0];
-      expect(typeof dto.error).toBe("string");
-      expect(dto.error.length).toBeGreaterThan(0);
-    });
-  });
-
-  it("skips non-prod levels when NODE_ENV=prod", () => {
-    process.env.NODE_ENV = "prod";
     jest.resetModules();
-    jest.isolateModules(() => {
-      const { LoggerService: ProdLogger } = require("../logger.service");
-      const req = makeReq();
-      ProdLogger.logInfo({ message: "m", req, res: makeRes(), service: "svc" });
-      ProdLogger.logDebug({ message: "m", req, service: "svc" });
-      expect(mockExecute).toHaveBeenCalledTimes(0);
-      ProdLogger.logError({
-        error: new Error("e"),
-        req,
-        res: makeRes(),
-        service: "svc",
+  });
+  describe("LoggerService 🧾", () => {
+    it("should create a logger instance", async () => {
+      const LoggerService = (await import("../logger.service")).default;
+
+      const loggerService = new LoggerService();
+      expect(loggerService).toBeInstanceOf(LoggerService);
+    });
+
+    it("should call addColors when create a logger instance", () => {
+      const winston = require("winston");
+      const LoggerService = require("../logger.service").default;
+
+      const addColorsSpy = jest.spyOn(winston, "addColors");
+
+      const loggerService = new LoggerService();
+      expect(loggerService).toBeInstanceOf(LoggerService);
+      expect(addColorsSpy).toHaveBeenCalledWith(LoggerService._logColors);
+    });
+
+    it("should obtain the default transporters using the envs", () => {
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const transporters = loggerService._getTransporters();
+
+      expect(transporters).toHaveLength(1);
+      expect(transporters[0]).toBeInstanceOf(transports.Console);
+    });
+
+    it("should obtain the file transporters using the envs", () => {
+      jest.mock("@/infrastructure/config/environment/envs", () => ({
+        LOG_TRANSPORT: "file",
+      }));
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const transporters = loggerService._getTransporters();
+
+      expect(transporters).toHaveLength(1);
+      expect(transporters[0]).toBeInstanceOf(transports.File);
+    });
+
+    it("should obtain the all transporters using the envs", () => {
+      jest.mock("@/infrastructure/config/environment/envs", () => ({
+        LOG_TRANSPORT: "all",
+      }));
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const transporters = loggerService._getTransporters();
+
+      expect(transporters).toHaveLength(2);
+      expect(transporters[0]).toBeInstanceOf(transports.Console);
+      expect(transporters[1]).toBeInstanceOf(transports.File);
+    });
+
+    it("should obtain the default log level using the envs", () => {
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const logLevel = loggerService._getLogLevel();
+
+      expect(logLevel).toBe("error");
+    });
+
+    it("should obtain the DEV environment logLevel using the envs", () => {
+      jest.mock("@/infrastructure/config/environment/envs", () => ({
+        NODE_ENV: "dev",
+      }));
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const logLevel = loggerService._getLogLevel();
+
+      expect(logLevel).toBe("silly");
+    });
+
+    it("should obtain the QA environment logLevel using the envs", () => {
+      jest.mock("@/infrastructure/config/environment/envs", () => ({
+        NODE_ENV: "qa",
+      }));
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const logLevel = loggerService._getLogLevel();
+
+      expect(logLevel).toBe("http");
+    });
+
+    it("should obtain the PROD environment logLevel using the envs", () => {
+      jest.mock("@/infrastructure/config/environment/envs", () => ({
+        NODE_ENV: "prod",
+      }));
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const logLevel = loggerService._getLogLevel();
+
+      expect(logLevel).toBe("info");
+    });
+
+    it("should obtain format for logs", () => {
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+      const logFormat = loggerService._getLogFormat();
+
+      expect(logFormat).toBeDefined();
+    });
+
+    it("should format logs as JSON", async () => {
+      const format = (await import("winston")).format;
+      const formatJsonSpy = jest.spyOn(format, "json");
+
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+
+      loggerService._getLogFormat();
+
+      expect(loggerService).toBeDefined();
+      expect(formatJsonSpy).toHaveBeenCalledTimes(1);
+      expect(formatJsonSpy).toHaveBeenCalledWith();
+    });
+
+    it("should format logs whit timestamp", async () => {
+      const format = (await import("winston")).format;
+      const formatTimeStampSpy = jest.spyOn(format, "timestamp");
+
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+
+      loggerService._getLogFormat();
+
+      expect(loggerService).toBeDefined();
+      expect(formatTimeStampSpy).toHaveBeenCalledTimes(1);
+      expect(formatTimeStampSpy).toHaveBeenCalledWith();
+    });
+
+    it("should format error logs with stack trace", async () => {
+      const format = (await import("winston")).format;
+      const formatErrorsSpy = jest.spyOn(format, "errors");
+
+      const LoggerService = require("../logger.service").default;
+      const loggerService = new LoggerService();
+
+      loggerService._getLogFormat();
+
+      expect(loggerService).toBeDefined();
+      expect(formatErrorsSpy).toHaveBeenCalledTimes(1);
+      expect(formatErrorsSpy).toHaveBeenCalledWith({
+        stack: true,
       });
-      ProdLogger.logWarn({
-        error: new Error("w"),
-        req,
-        res: makeRes(),
-        service: "svc",
-      });
-      expect(mockExecute).toHaveBeenCalledTimes(2);
     });
   });
 });
